--- src/Library/Data/Core/Data.Core/Internal/CommonExtensions.cs
@@ -1,75 +1,0 @@
-﻿using System;
-
-namespace Nm.Lib.Data.Core.Internal
-{
-    /// <summary>
-    /// 通用扩展方法
-    /// </summary>
-    internal static class CommonExtensions
-    {
-        /// <summary>
-        /// 转换成Int/Int32
-        /// </summary>
-        /// <param name="s"></param>
-        /// <param name="round">是否四舍五入，默认false</param>
-        /// <returns></returns>
-        public static int ToInt(this object s, bool round = false)
-        {
-            if (s == null || s == DBNull.Value)
-                return 0;
-
-            if (s.GetType().IsEnum)
-            {
-                return (int)s;
-            }
-
-            if (s is bool b)
-                return b ? 1 : 0;
-
-            if (int.TryParse(s.ToString(), out int result))
-                return result;
-
-            var f = s.ToFloat();
-            return round ? Convert.ToInt32(f) : (int)f;
-        }
-
-        /// <summary>
-        /// 转换成Float/Single
-        /// </summary>
-        /// <param name="s"></param>
-        /// <returns></returns>
-        public static float ToFloat(this object s)
-        {
-            if (s == null || s == DBNull.Value)
-                return 0f;
-
-            float.TryParse(s.ToString(), out float result);
-            return result;
-        }
-
-        /// <summary>
-        /// 转换成DateTime
-        /// </summary>
-        /// <param name="s"></param>
-        /// <returns></returns>
-        public static DateTime ToDateTime(this object s)
-        {
-            if (s == null || s == DBNull.Value)
-                return DateTime.MinValue;
-
-            DateTime.TryParse(s.ToString(), out DateTime result);
-            return result;
-        }
-
-        /// <summary>
-        /// 泛型转换，转换失败会抛出异常
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="s"></param>
-        /// <returns></returns>
-        public static T To<T>(this object s)
-        {
-            return (T)Convert.ChangeType(s, typeof(T));
-        }
-    }
-}
