--- src/Library/Data/Core/Data.Core/Internal/StringExtensions.cs
@@ -1,25 +1,0 @@
-﻿namespace Nm.Lib.Data.Core.Internal
-{
-    internal static class StringExtensions
-    {
-        /// <summary>
-        /// 判断字符串是否为Null、空
-        /// </summary>
-        /// <param name="s"></param>
-        /// <returns></returns>
-        public static bool IsNull(this string s)
-        {
-            return string.IsNullOrWhiteSpace(s);
-        }
-
-        /// <summary>
-        /// 判断字符串是否不为Null、空
-        /// </summary>
-        /// <param name="s"></param>
-        /// <returns></returns>
-        public static bool NotNull(this string s)
-        {
-            return !string.IsNullOrWhiteSpace(s);
-        }
-    }
-}
