--- src/Library/Data/Core/Data.Core/Internal/TypeExtensions.cs
@@ -1,48 +1,0 @@
-﻿using System;
-using System.Reflection;
-
-namespace Nm.Lib.Data.Core.Internal
-{
-    /// <summary>
-    /// 类型扩展
-    /// </summary>
-    internal static class TypeExtensions
-    {
-        /// <summary>
-        /// 判断属性是否是静态的
-        /// </summary>
-        /// <param name="property"></param>
-        /// <returns></returns>
-        public static bool IsStatic(this PropertyInfo property) => (property.GetMethod ?? property.SetMethod).IsStatic;
-
-        /// <summary>
-        /// 判断是否继承自指定的泛型
-        /// </summary>
-        /// <param name="type"></param>
-        /// <param name="generic"></param>
-        /// <returns></returns>
-        public static bool IsSubclassOfGeneric(this Type type, Type generic)
-        {
-            while (type != null && type != typeof(object))
-            {
-                var cur = type.IsGenericType ? type.GetGenericTypeDefinition() : type;
-                if (generic == cur)
-                {
-                    return true;
-                }
-                type = type.BaseType;
-            }
-            return false;
-        }
-
-        /// <summary>
-        /// 判断是否可空类型
-        /// </summary>
-        /// <param name="type"></param>
-        /// <returns></returns>
-        public static bool IsNullable(this Type type)
-        {
-            return type.IsGenericType && type.GetGenericTypeDefinition() == typeof(Nullable<>);
-        }
-    }
-}
