+++ src/Library/Data/Core/Data.Core/SqlQueryable/Internal/QueryBody.cs
@@ -6,7 +6,8 @@ using System.Linq.Expressions;
 using Nm.Lib.Data.Abstractions;
 using Nm.Lib.Data.Abstractions.Enums;
 using Nm.Lib.Data.Abstractions.Pagination;
-using Nm.Lib.Data.Core.Internal;
+using Nm.Lib.Utils.Core;
+using Nm.Lib.Utils.Core.Extensions;
 
 namespace Nm.Lib.Data.Core.SqlQueryable.Internal
 {
