+++ src/Library/Data/Core/Data.Core/SqlQueryable/NetSqlQueryable4.cs
@@ -9,9 +9,9 @@ using Nm.Lib.Data.Abstractions.Enums;
 using Nm.Lib.Data.Abstractions.Pagination;
 using Nm.Lib.Data.Abstractions.SqlQueryable;
 using Nm.Lib.Data.Abstractions.SqlQueryable.GroupByQueryable;
-using Nm.Lib.Data.Core.Internal;
 using Nm.Lib.Data.Core.SqlQueryable.GroupByQueryable;
 using Nm.Lib.Data.Core.SqlQueryable.Internal;
+using Nm.Lib.Utils.Core;
 
 namespace Nm.Lib.Data.Core.SqlQueryable
 {
