+++ src/Library/Data/Db/Data.SQLite/SQLiteAdapter.cs
@@ -2,7 +2,7 @@
 using System.Text;
 using Nm.Lib.Data.Abstractions.Enums;
 using Nm.Lib.Data.Core;
-using Nm.Lib.Data.Core.Internal;
+using Nm.Lib.Utils.Core.Helpers;
 
 namespace Nm.Lib.Data.SQLite
 {
