+++ src/Library/Utils/Utils.Core/Extensions/TypeExtensions.cs
@@ -79,5 +79,16 @@ namespace Nm.Lib.Utils.Core.Extensions
             }
             return false;
         }
+
+
+        /// <summary>
+        /// 判断是否可空类型
+        /// </summary>
+        /// <param name="type"></param>
+        /// <returns></returns>
+        public static bool IsNullable(this Type type)
+        {
+            return type.IsGenericType && type.GetGenericTypeDefinition() == typeof(Nullable<>);
+        }
     }
 }
